// frontend/src/api/models/DbTask.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Rankode
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { PgtypeTimestamp } from './PgtypeTimestamp';
import {
    PgtypeTimestampFromJSON,
    PgtypeTimestampFromJSONTyped,
    PgtypeTimestampToJSON,
    PgtypeTimestampToJSONTyped,
} from './PgtypeTimestamp';

/**
 * 
 * @export
 * @interface DbTask
 */
export interface DbTask {
    /**
     * 
     * @type {PgtypeTimestamp}
     * @memberof DbTask
     */
    createdAt?: PgtypeTimestamp;
    /**
     * 
     * @type {string}
     * @memberof DbTask
     */
    description?: string;
    /**
     * 
     * @type {number}
     * @memberof DbTask
     */
    difficulty?: number;
    /**
     * 
     * @type {number}
     * @memberof DbTask
     */
    id?: number;
    /**
     * 
     * @type {number}
     * @memberof DbTask
     */
    passes?: number;
    /**
     * 
     * @type {number}
     * @memberof DbTask
     */
    score?: number;
    /**
     * 
     * @type {string}
     * @memberof DbTask
     */
    title?: string;
    /**
     * 
     * @type {Array<number>}
     * @memberof DbTask
     */
    topics?: Array<number>;
    /**
     * 
     * @type {PgtypeTimestamp}
     * @memberof DbTask
     */
    updatedAt?: PgtypeTimestamp;
    /**
     * 
     * @type {number}
     * @memberof DbTask
     */
    userId?: number;
}

/**
 * Check if a given object implements the DbTask interface.
 */
export function instanceOfDbTask(value: object): value is DbTask {
    return true;
}

export function DbTaskFromJSON(json: any): DbTask {
    return DbTaskFromJSONTyped(json, false);
}

export function DbTaskFromJSONTyped(json: any, ignoreDiscriminator: boolean): DbTask {
    if (json == null) {
        return json;
    }
    return {
        
        'createdAt': json['created_at'] == null ? undefined : PgtypeTimestampFromJSON(json['created_at']),
        'description': json['description'] == null ? undefined : json['description'],
        'difficulty': json['difficulty'] == null ? undefined : json['difficulty'],
        'id': json['id'] == null ? undefined : json['id'],
        'passes': json['passes'] == null ? undefined : json['passes'],
        'score': json['score'] == null ? undefined : json['score'],
        'title': json['title'] == null ? undefined : json['title'],
        'topics': json['topics'] == null ? undefined : json['topics'],
        'updatedAt': json['updated_at'] == null ? undefined : PgtypeTimestampFromJSON(json['updated_at']),
        'userId': json['user_id'] == null ? undefined : json['user_id'],
    };
}

export function DbTaskToJSON(json: any): DbTask {
    return DbTaskToJSONTyped(json, false);
}

export function DbTaskToJSONTyped(value?: DbTask | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'created_at': PgtypeTimestampToJSON(value['createdAt']),
        'description': value['description'],
        'difficulty': value['difficulty'],
        'id': value['id'],
        'passes': value['passes'],
        'score': value['score'],
        'title': value['title'],
        'topics': value['topics'],
        'updated_at': PgtypeTimestampToJSON(value['updatedAt']),
        'user_id': value['userId'],
    };
}