// frontend/src/api/models/DbTopic.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Rankode
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { PgtypeTimestamp } from './PgtypeTimestamp';
import {
    PgtypeTimestampFromJSON,
    PgtypeTimestampFromJSONTyped,
    PgtypeTimestampToJSON,
    PgtypeTimestampToJSONTyped,
} from './PgtypeTimestamp';

/**
 * 
 * @export
 * @interface DbTopic
 */
export interface DbTopic {
    /**
     * 
     * @type {PgtypeTimestamp}
     * @memberof DbTopic
     */
    createdAt?: PgtypeTimestamp;
    /**
     * 
     * @type {number}
     * @memberof DbTopic
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof DbTopic
     */
    name?: string;
    /**
     * 
     * @type {number}
     * @memberof DbTopic
     */
    tasksCount?: number;
}

/**
 * Check if a given object implements the DbTopic interface.
 */
export function instanceOfDbTopic(value: object): value is DbTopic {
    return true;
}

export function DbTopicFromJSON(json: any): DbTopic {
    return DbTopicFromJSONTyped(json, false);
}

export function DbTopicFromJSONTyped(json: any, ignoreDiscriminator: boolean): DbTopic {
    if (json == null) {
        return json;
    }
    return {
        
        'createdAt': json['created_at'] == null ? undefined : PgtypeTimestampFromJSON(json['created_at']),
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
        'tasksCount': json['tasks_count'] == null ? undefined : json['tasks_count'],
    };
}

export function DbTopicToJSON(json: any): DbTopic {
    return DbTopicToJSONTyped(json, false);
}

export function DbTopicToJSONTyped(value?: DbTopic | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'created_at': PgtypeTimestampToJSON(value['createdAt']),
        'id': value['id'],
        'name': value['name'],
        'tasks_count': value['tasksCount'],
    };
}